Split ShouldlyConfiguration into public + trip-wire partial files
ShouldlyConfiguration.cs keeps the public surface
(DisableSourceInErrors, IsSourceDisabledInErrors, DefaultTaskTimeout,
DiffStyle, EscapeStyle, CompareAsObjectTypes).

ShouldlyConfiguration.TripWire.cs holds the internal test-infrastructure
helpers (AssertCallerArgumentExpressionIsUsed, AllowStackWalking,
IsCallerArgumentExpressionRequired) so the public file isn't cluttered
with them.
